fullstackvideo/requirements.txt
asgiref==3.5.2
boto3==1.26.6
botocore==1.29.6
confluent-kafka==2.3.0
Django==4.1.3
django-cors-headers==3.13.0
django-storages==1.13.1
djangorestframework==3.14.0
ffmpeg-python==0.2.0
jmespath==1.0.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
